+++ lib/src/types/main.dart
@@ -223,3 +223,5 @@ export 'activity_button.dart' show ActivityButton;
 export 'ui_event_attribution.dart' show UIEventAttribution;
 export 'tracing_mode.dart' show TracingMode;
 export 'tracing_category.dart' show TracingCategory;
+export 'custom_tabs_post_message_result_type.dart'
+    show CustomTabsPostMessageResultType;