+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 6.0.0-beta.27
+
+- Added `requestPostMessageChannel`, `postMessage`, `isEngagementSignalsApiAvailable` methods on `ChromeSafariBrowser` for Android
+- Added `onMessageChannelReady`, `onPostMessage`, `onVerticalScrollEvent`, `onGreatestScrollPercentageIncreased`, `onSessionEnded` events on `ChromeSafariBrowser` for Android
+
 ## 6.0.0-beta.26
 
 - Throw an error if any controller is used after being disposed